fix: fixed int128 conversion to unit256

+++ evm/src/curve/CurveAdapter.sol
@@ -256,8 +256,8 @@ contract CurveAdapter is ISwapAdapter {
         returns (Fraction memory)
     {
         uint256 amountIn;
-        uint256 sellTokenIndexUint = uint256(sellParams.sellTokenIndex);
-        uint256 buyTokenIndexUint = uint256(sellParams.buyTokenIndex);
+        uint256 sellTokenIndexUint = uint256(uint128(sellParams.sellTokenIndex));
+        uint256 buyTokenIndexUint = uint256(uint128(sellParams.buyTokenIndex));
 
         if (sellParams.isSwappingUnderlying) {
             amountIn = ICurveStableSwapMetaPool(sellParams.poolAddress).balances(
